Fix CutsceneOamData type, match few more functions

+++ include/structs/cutscene.h
@@ -124,11 +124,11 @@ struct CutsceneOamData {
     u8 animationDurationCounter;
     u8 currentAnimationFrame;
     u8 oamID;
-    u8 priority:2;
-    u8 unk_B_4:2;
-    u8 ended:1;
-    u8 notDrawn:1;
-    u8 idChanged:2;
+    u16 priority:2;
+    u16 unk_B_4:2;
+    u16 ended:1;
+    u16 notDrawn:1;
+    u16 idChanged:2;
     u8 unk_C_0:4;
     u8 rotationScaling:1;
     u8 unk_C_5:1;
